third_party/WebKit/LayoutTests/http/tests/security/javascriptURL/javascriptURL-execution-context-frame-location-htmldom.html

<html>
<head>
    <script>
        if (window.testRunner) {
            testRunner.dumpAsText();
            testRunner.dumpChildFramesAsText();
            testRunner.waitUntilDone();
        }

        window.addEventListener('message', function() {
            runTest();
            if (window.testRunner)
                testRunner.notifyDone();
        });
        
        setFrameLocation = function(url) {
            var frame = document.getElementById('aFrame');
            var jsErrorMessage = 'Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.';
            try {
                frame.location = url;
            } catch (e) {
                console.log("Caught exception while setting frame's location to '" + url + "'. '" + e + "'.");
                if (e.message == jsErrorMessage)
                    console.log("PASS: Exception was '" + e.message + "'.");
                else
                    console.log("FAIL: Exception should have been '" + jsErrorMessage + "', was '" + e.message + "'.");
            }
        }

        function runTest() {
            setFrameLocation('javascript:"FAIL: this should not have been loaded."');
            setFrameLocation(' javascript:"FAIL: this should not have been loaded."');
            setFrameLocation('java\0script:"FAIL: this should not have been loaded."');
            setFrameLocation('javascript\t:"FAIL: this should not have been loaded."');
            setFrameLocation('javascript\1:"FAIL: this should not have been loaded."');
        }
    </script>
</head>
<frameset rows="10%,*">
    <frame id='main' src="data:text/html,<p>This test passes as long as the word 'FAIL' and an accompanying message do not appear in the frame below.</p>">
    <frame id='aFrame' src='http://localhost:8000/security/resources/cross-frame-iframe.html'>
</frameset>
</html>
