third_party/WebKit/LayoutTests/http/tests/security/javascriptURL/javascriptURL-execution-context-frame-location-htmldom-expected.txt

CONSOLE MESSAGE: line 19: Caught exception while setting frame's location to 'javascript:"FAIL: this should not have been loaded."'. 'SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.'.
CONSOLE MESSAGE: line 21: PASS: Exception was 'Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.'.
CONSOLE MESSAGE: line 19: Caught exception while setting frame's location to ' javascript:"FAIL: this should not have been loaded."'. 'SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.'.
CONSOLE MESSAGE: line 21: PASS: Exception was 'Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.'.
CONSOLE MESSAGE: line 19: Caught exception while setting frame's location to 'javascript	:"FAIL: this should not have been loaded."'. 'SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.'.
CONSOLE MESSAGE: line 21: PASS: Exception was 'Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.'.


--------
Frame: 'main'
--------
This test passes as long as the word 'FAIL' and an accompanying message do not appear in the frame below.

--------
Frame: 'aFrame'
--------
Inner iframe.
